requirements.txt
requests
pandas
python-dotenv
pytz
tzdata
